+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Types of changes:
 - `<Table/>` - fix broken `<TableToolbar/>` docs and `<TableActionCell/>` RTL [#2224](https://github.com/wix/wix-style-react/pull/2224)
 - `<SideMenu/>` - support `dataHook` prop [#2256](https://github.com/wix/wix-style-react/pull/2256)
 - Polyfills - fix `raf` not setting all methods on the global variable [#2258](https://github.com/wix/wix-style-react/pull/2258)
+- `Tooltip` - remove `stopPropagation` from `onClick` as it breaks some use cases [#2260](https://github.com/wix/wix-style-react/pull/2260)
 
 ## 5.3.2 - 2018-09-20