+++ CHANGELOG.md
@@ -14,12 +14,18 @@ Types of changes:
 1. **Docs** for documentation changes
 
 ## Next
+
+### Changed
+
+- `<InputWithTags/>` - fix focus background color [#2245](https://github.com/wix/wix-style-react/pull/2245)
+
 ### Fixed
 - `<Calendar/>` - remove shadow and border-radius from component [#2205](https://github.com/wix/wix-style-react/pull/2205)
 - `<Tooltip/>` - add popover prop [#2205](https://github.com/wix/wix-style-react/pull/2205)
 - `<Table/>` - fix broken `<TableToolbar/>` docs and `<TableActionCell/>` RTL [#2224](https://github.com/wix/wix-style-react/pull/2224)
 - `<SideMenu/>` - support `dataHook` prop [#2256](https://github.com/wix/wix-style-react/pull/2256)
 - Polyfills - fix `raf` not setting all methods on the global variable [#2258](https://github.com/wix/wix-style-react/pull/2258)
+- `Tooltip` - remove `stopPropagation` from `onClick` as it breaks some use cases [#2260](https://github.com/wix/wix-style-react/pull/2260)
 
 ## 5.3.2 - 2018-09-20